CT-1518 add policy for CasesController#show, restricting SAR access
Unlike FOI cases, SAR cannot be viewed by just anyone for privacy
considerations. This PR adds a policy for the show action and restricts
visibility of SAR cases to managers or responders who are working on the case
... OR who are working on a case that is linked to the case being viewed, to
allow for TMMs or case reviews.
