+++ app/controllers/cases_controller.rb
@@ -177,6 +177,8 @@ class CasesController < ApplicationController
     if policy(@case).can_accept_or_reject_responder_assignment?
       redirect_to edit_case_assignment_path @case, @case.responder_assignment.id
     else
+      authorize @case
+
       get_flash_errors_for_case(@case)
       set_permitted_events
       @accepted_now = params[:accepted_now]
